% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoIGV.R
\name{AutoIGV}
\alias{AutoIGV}
\title{Function to produce IGV session files and screenshorts automatically}
\usage{
AutoIGV(
  coords,
  indents = c(0, 0),
  bams.paths,
  refs.paths = data.table(paths = ""[0], name = ""[0]),
  max.matrix = data.table(),
  folder.name = tempdir()
)
}
\arguments{
\item{coords}{data frame with 6 columns: chr, start, end, maximum, 
coverage.only and suffix.to.file. Maximum is the maximum of the
coverage tracks to be displayd. If you want to use auto maximum, 
put -1 to maximum column. Column coverage.only indicates,
whatever or not only coverage tracks should be displayed: FALSE, if you want
reads to be displayed too and TRUE otherwise. Column 
suffix.to.file contains suffixes to be appended to files names}

\item{indents}{vector of the nuumber of nucleoteds to be dysplayed from the
left and rigth of region of interest. Default value is c(0, 0)}

\item{bams.paths}{vector, containing paths to bam files. The order 
matters: tracks will appear in igv according to the order of 
this vector}

\item{refs.paths}{vector, containing paths to reference files. The order
also matters}

\item{folder.name}{name of the folder to put results in}

\item{coverage.only}{boolean, indicating, whatever or not only coverage 
tracks will be diplayed}
}
\value{
folder with one IGV session and 1 png per line of coords data 
        frame
}
\description{
Function to produce IGV session files and screenshorts automatically
}
\author{
Maria Litovchenko
}
