% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{copyODS}
\alias{copyODS}
\title{Make a clean copy of the original ODS}
\usage{
copyODS(ods, keep.meta = T)
}
\arguments{
\item{ods}{OUTRIDER Dataset or RangedSummarizedExperiment}

\item{keep.meta}{inf TRUE keep mcols, otherwise remove (default FALSE)}
}
\value{
copy of ods with 2 assays "counts" and "trueCounts"
}
\description{
Removes old assays and keeps only original counts
Removes any meta data
}
