% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{coverage_ratio}
\alias{coverage_ratio}
\title{for checking the coverage}
\usage{
coverage_ratio(exons_dt)
}
\arguments{
\item{exons_dt}{data.table containing the counts and summary values as returned by summarise_counts()}
}
\description{
for checking the coverage
}
