% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_exploration.R
\name{coverageMatrix}
\alias{coverageMatrix}
\title{Coverage Matrix}
\usage{
coverageMatrix(cvg_list, regions, strand, feature, window, parallel = F)
}
\arguments{
\item{cvg_list}{list of coverage Rles}

\item{regions}{data.table containing the feature regions, must contain a column "site"}

\item{strand}{either "+" or "-"}

\item{feature}{type of feature: "TSS" or "PAS"}

\item{window}{number of positions upstream and downstream of feature site}
}
\description{
Generates a matrix with average per-base counts around TSS or PAS of given transcripts
}
