% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_exploration.R
\name{coverageMatrixPerSite}
\alias{coverageMatrixPerSite}
\title{Coverage Matrix per Site}
\usage{
coverageMatrixPerSite(cvg_list, regions, gene, feature = "TSS", window = 500)
}
\arguments{
\item{cvg_list}{a named list of coverage Rle's}

\item{regions}{data.table containing the feature regions, must contain a column "site"}

\item{gene}{name of gene}

\item{feature}{type of feature: "TSS" or "PAS"}

\item{window}{number of positions upstream and downstream of feature site}

\item{strand}{either "+" or "-"}
}
\description{
Generates a matrix with per-base counts around a specific TSS or PAS for all samples
}
