% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_exploration.R
\name{drawCoverageHeatmap}
\alias{drawCoverageHeatmap}
\title{Draw a heatmap}
\usage{
drawCoverageHeatmap(cvg_matrix, strand, feature, window, gene = NULL, log)
}
\arguments{
\item{strand}{either "+" or "-"}

\item{feature}{either "TSS" or "PAS"}

\item{window}{positions upstream/downstream}

\item{log}{TRUE or FALSE for log10-transformation of counts + 1}

\item{cvg}{coverage matrix}

\item{transcript}{ID of the transcript to be displayed in the plot for single-Transcript plots}
}
\value{
a heatmap depending on the input (per site if transcript is specified or for multiple transcripts).
}
\description{
Wraps the actual plotting, only dependent on matrix and meta info
}
