% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{exon_to_tx}
\alias{exon_to_tx}
\title{maps the exons to the transcripts}
\usage{
exon_to_tx(txdb)
}
\arguments{
\item{txdb}{annotation database}
}
\value{

}
\description{
maps the exons to the transcripts
}
