% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{extend}
\alias{extend}
\title{Extend regions upstream or downstream}
\usage{
extend(gr, direction, window)
}
\arguments{
\item{gr}{regions as GenomicRanges}

\item{direction}{upstream or downstream}

\item{window}{size of extension}
}
\description{
Extend regions upstream or downstream
}
