% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{featureInfo}
\alias{featureInfo}
\title{Retrieve info as data.table}
\usage{
featureInfo(gr, columns = NULL, summary = F, rownumbers = F)
}
\arguments{
\item{gr}{GRanges object or any object containing it (i.e. RangedSummarizedExperiment, OutriderDataSet), must support mcols function}

\item{columns}{a character vector for selecting only a part of the mcols. If NULL then all the columns will be included}

\item{summary}{include summary per feature over all samples, if TRUE}

\item{rownumbers}{if TRUE, the row numbers will be included as columns "row"}
}
\description{
Get data.table of metadata from RangedSummarizedExperiment or OutriderDataSet object
}
