% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{generalise_gene_type}
\alias{generalise_gene_type}
\title{Summarise gene type}
\usage{
generalise_gene_type(
  dt,
  mapping = NULL,
  column = "gene_type",
  protein_coding = "protein_coding",
  general_column = "general_gene_type"
)
}
\arguments{
\item{dt}{data.table with at least a column containing specific gene types}

\item{mapping}{mapping of specific to general gene type as data.table or list of vectors
if kept NULL, then a basic mapping of pc (protein_coding) vs. npc is created}

\item{column}{specifies the column of the specific gene type}

\item{general_column}{name of general gene type column to add}
}
\description{
Add a column named general_gene_type to the data.table containing generalised gene type names
}
