% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{get_last_exons}
\alias{get_last_exons}
\title{simple function that takes the first hit after sorting}
\usage{
get_last_exons(dt, by)
}
\arguments{
\item{dt}{containing exon info}

\item{by}{sorting criterium (transcript or gene)}
}
\description{
simple function that takes the first hit after sorting
}
