% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{getDassieRegions}
\alias{getDassieRegions}
\title{Extract regions around TSS and PAS and returns a GRanges object}
\usage{
getDassieRegions(
  exon_dt,
  tssWindow = 500,
  pasWindow = 1000,
  tx_id = "tx_id",
  remove_redundant = TRUE
)
}
\arguments{
\item{exon_dt}{data.table convertible into GRanges, and contain a transcript
column that is specified in tx_id. Data must be strand-specific.}

\item{window}{number of bp that the regions should have}
}
\description{
Convenience function for TSS and PAS regions of same length
}
