% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{getExons}
\alias{getExons}
\title{Get exons from txdb}
\usage{
getExons(txdb, filter_length = "tx_length", filter_exon = TRUE)
}
\arguments{
\item{filter_length}{filter transcripts
'tx_length': by length of transcript
'gene_length': by length of gene}

\item{filter_exon}{keep transcript with longest exons}
}
\value{
exons in a data.table convertible to GRanges object, contains
        transcript and gene mapping
}
\description{
Get exons from txdb
}
