% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{gr_to_bed}
\alias{gr_to_bed}
\title{converts GRanges object to bed file}
\usage{
gr_to_bed(gr, filename, invertStrand = F)
}
\arguments{
\item{gr}{GRanges object}

\item{filename}{for output bed file}

\item{invert}{strand if TRUE}
}
\description{
converts GRanges object to bed file
}
