% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{injectFeatures}
\alias{injectFeatures}
\title{Take random features and replace these in simulated}
\usage{
injectFeatures(ods, shifted, prob = 1e-04, seed = NULL)
}
\arguments{
\item{ods}{OutriderDataSet containing real counts}

\item{shifted}{RangedSummarizedExperiment on shifted features}

\item{prob}{probability of an outlier}

\item{seed}{for sampling the positions of injection}

\item{n}{number of features to randomly choose}
}
\value{
simulated OutriderDataSet containing outliers at random positions and INDICES of injections
}
\description{
Take random features and replace these in simulated
}
