% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{injectSample}
\alias{injectSample}
\title{Inject a complete sample from shifted ods into original ods}
\usage{
injectSample(ods, shifted, sampleID)
}
\arguments{
\item{ods}{OutriderDataSet containing real counts}

\item{shifted}{RangedSummarizedExperiment on shifted features}

\item{sampleID}{ID of sample (patient)}
}
\value{
simulated OutriderDataSet containing all simulated outliers for 1 sample
}
\description{
Inject a complete sample from shifted ods into original ods
}
