% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{multiplotLoop}
\alias{multiplotLoop}
\title{Plot multiple results with a loop}
\usage{
multiplotLoop(ods, ods_gene, geneName, feature_dt = NULL, plot = plotQQ, ...)
}
\arguments{
\item{ods}{Outrider dataset}

\item{geneName}{name of gene to be aggregated over}

\item{feature_dt}{dt containing feature information, must be subsettable by gene_name}

\item{plot}{a plot function with params (ods, geneID)}

\item{export}{}
}
\description{
Plot multiple results with a loop
}
