% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{pasRegions}
\alias{pasRegions}
\title{Polyadenylation site regions}
\usage{
pasRegions(exon_dt, window, tx_id = "tx_id")
}
\arguments{
\item{window}{number of bp for regions}

\item{exons_dt}{data.table of GRanges object with columns "tx_id" for each
exon for merging by transcripts. Ranges must be strand-specific.}
}
\value{
data.table with coordinates and some metadata on the new regions
exon_id and transcript_id are necessary for joining the features with the rest of the exon annotation.
}
\description{
Get regions of specified length around the PAS
}
