% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_exon_distribution}
\alias{plot_exon_distribution}
\title{plots number of exons according to gene type}
\usage{
plot_exon_distribution(
  exons,
  select = NULL,
  title = "Number of exons by gene type",
  filename = NA,
  ...
)
}
\arguments{
\item{exons}{must be a data.table with genetic features in rows and samples in columns
must also contain at least gene type}

\item{select}{gene types}

\item{title}{displayed in plot}

\item{filename}{if not specified, the plot will not be saved}
}
\description{
plots number of exons according to gene type
}
