% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mean_counts}
\alias{plot_mean_counts}
\title{Boxplots the distribution of mean counts per gene compared between gene types}
\usage{
plot_mean_counts(
  exons,
  select = NULL,
  title = "Mean exon counts",
  filename = NA,
  ...
)
}
\arguments{
\item{exons}{must be a data.table with genetic features in rows and samples in columns
must also contain at least gene type}

\item{select}{gene types}

\item{title}{displayed in plot}

\item{filename}{if not specified, the plot will not be saved}
}
\description{
Boxplots the distribution of mean counts per gene compared between gene types
}
