% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_exploration.R
\name{plotCoverageHeatmap}
\alias{plotCoverageHeatmap}
\title{Visualise coverage for multiple genes using average counts}
\usage{
plotCoverageHeatmap(
  cvg_list,
  regions,
  strand = "+",
  feature = "TSS",
  window = 500,
  log = T,
  parallel = F
)
}
\arguments{
\item{cvg_list}{a named list of coverage Rle's}

\item{regions}{GRanges object containing the feature regions, must contain a column "site"}

\item{strand}{either "+" or "-"}

\item{feature}{type of feature: "TSS" or "PAS"}

\item{window}{number of positions upstream and downstream of feature site}

\item{log}{log representation of counts}
}
\description{
Visualise coverage for multiple genes using average counts
}
