% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotExpressionRank2}
\alias{plotExpressionRank2}
\title{Expression Ranks with marked sample}
\usage{
plotExpressionRank2(
  ods,
  id,
  sampleIDs = c(),
  padj = 0.05,
  normalized = T,
  main = NULL,
  log = T
)
}
\arguments{
\item{ods}{OUTRIDER data set}

\item{sampleIDs}{sample IDs to be additionally marked, if not already marked signif}

\item{padj}{cutoff for marking most significant}

\item{normalized}{if T then normalized counts will be used}

\item{gene_id}{ID of ods features}
}
\description{
Expression Ranks with marked sample
}
