% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_exploration.R
\name{plotPerSiteCoverageHeatmap}
\alias{plotPerSiteCoverageHeatmap}
\title{Coverage Heatmap per site}
\usage{
plotPerSiteCoverageHeatmap(
  cvg_matrix,
  regions,
  gene,
  feature = "TSS",
  window = 500,
  log = T
)
}
\arguments{
\item{cvg_matrix}{matrix to be plotted (created by coverMatrixPerSite)}

\item{regions}{data.table containing the feature regions, must contain a column "site"}

\item{gene}{gene name}

\item{feature}{type of feature: "TSS" or "PAS"}

\item{window}{number of positions upstream and downstream of feature site}

\item{log}{log representation of counts}

\item{strand}{either "+" or "-"}
}
\description{
Visualise coverage across all samples for a given gene name
}
