% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotQQ2}
\alias{plotQQ2}
\title{QQ plots with marked samples}
\usage{
plotQQ2(ods, id, sampleIDs = c(), padj = 0.05, main = NULL, ci = 0.95)
}
\arguments{
\item{ods}{OUTRIDER data set}

\item{sampleIDs}{sample IDs to be additionally marked, if not already marked signif}

\item{padj}{cutoff for marking most significant}

\item{gene_id}{ID of ods features}
}
\description{
QQ plots with marked samples
}
