% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoIGV.R
\name{PrintAnnotationHeader}
\alias{PrintAnnotationHeader}
\title{PrintAnnotationHeader
Prints header of the annotation part to the session file.}
\usage{
PrintAnnotationHeader(file.path, heigth)
}
\arguments{
\item{file.path}{path to the session file to print in}

\item{heigth}{heigth of the track}
}
\description{
PrintAnnotationHeader
Prints header of the annotation part to the session file.
}
\author{
Maria Litovchenko
}
