% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoIGV.R
\name{PrintAnnotationTrack}
\alias{PrintAnnotationTrack}
\title{PrintAnnotationTrack
Prints annotation tracks to the session file}
\usage{
PrintAnnotationTrack(path.to.ann, file.path)
}
\arguments{
\item{path.to.ann}{path to the annotation file}

\item{file.path}{path to the session file to print in}
}
\description{
PrintAnnotationTrack
Prints annotation tracks to the session file
}
\author{
Maria Litovchenko
}
