% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoIGV.R
\name{PrintIGVbatch}
\alias{PrintIGVbatch}
\title{Prints IGV batch script file for further png profuction}
\usage{
PrintIGVbatch(session.path, bams.paths, batch.path)
}
\arguments{
\item{bams.paths}{paths to the bams files}

\item{batch.path}{path to the batch file to print in}
}
\value{
IGV batch template file
}
\description{
Prints IGV batch script file for further png profuction
}
\seealso{

}
\author{
Maria Litovchenko
}
\keyword{benchmark}
