% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoIGV.R
\name{PrintIGVSessionHeader}
\alias{PrintIGVSessionHeader}
\title{PrintIGVSessionHeader
Prints session header with coordinates and genome to the session file}
\usage{
PrintIGVSessionHeader(coords, indents = c(0, 0), file.path)
}
\arguments{
\item{coords}{vector with 5 items: chr, start, end, coverage.only,
maximum and suffix.to.file}

\item{indents}{vector of the number of nucleoteds to be dysplayed from the
left and rigth of region of interest. Default value is c(0, 0)}

\item{file.path}{path to the session file to print in}
}
\description{
PrintIGVSessionHeader
Prints session header with coordinates and genome to the session file
}
\author{
Maria Litovchenko
}
