% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoIGV.R
\name{PrintIGVSessionResourses}
\alias{PrintIGVSessionResourses}
\title{PrintIGVSessionResourses
Prints session resourses with paths to files}
\usage{
PrintIGVSessionResourses(paths.to.files, file.path)
}
\arguments{
\item{paths.to.files}{vector, containing paths to bam files. 
The order matters: tracks will appear in igv according to the 
order of this vector}

\item{file.path}{path to the session file to print in}
}
\description{
PrintIGVSessionResourses
Prints session resourses with paths to files
}
\author{
Maria Litovchenko
}
