% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoIGV.R
\name{PrintIGVTrack}
\alias{PrintIGVTrack}
\title{PrintIGVTrack
Prints tracks of bams to the session file}
\usage{
PrintIGVTrack(bams.paths, coverage.only, file.path)
}
\arguments{
\item{coverage.only}{parameter, which equals to -1, if reads are also
displayed and the heigth of the track otherrwise}

\item{file.path}{path to the session file to print in}

\item{bam.paths}{path to the bam file}
}
\description{
PrintIGVTrack
Prints tracks of bams to the session file
}
\author{
Maria Litovchenko
}
