% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoIGV.R
\name{ProduceIGVSession}
\alias{ProduceIGVSession}
\title{ProduceIGVSession
Produces session file for 1 entry of coords data frame}
\usage{
ProduceIGVSession(
  region.params,
  indents = c(0, 0),
  bams.paths,
  refs.paths = data.table(paths = ""[0], name = ""[0]),
  max.matrix = data.table(),
  dir.path = tempdir()
)
}
\arguments{
\item{region.params}{row of coords data frame}

\item{indents}{vector of the number of nucleotids to be displayed from
the left and rigth of region of interest. Default value is 
c(0, 0)}

\item{bams.paths}{vector, containing paths to bam files. The order 
matters: tracks will appear in igv according to the order
of this vector}

\item{refs.paths}{vector, containing paths to reference files. 
The order also matters.}

\item{dir.path}{name of the folder to put results in}
}
\value{
folder with one IGV session and 1 png per line of coords data frame
}
\description{
ProduceIGVSession
Produces session file for 1 entry of coords data frame
}
\author{
Maria Litovchenko
}
