% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{removeZeroCounts}
\alias{removeZeroCounts}
\title{Remove simulated 0-count rows}
\usage{
removeZeroCounts(ods)
}
\arguments{
\item{ods}{simulated OutriderDataSet containing injections}
}
\description{
Remove simulated 0-count rows
}
