% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoIGV.R
\name{set_AutoIGV_DEBUG}
\alias{set_AutoIGV_DEBUG}
\title{toogles the debuging mode for AutoIGV on and off}
\usage{
set_AutoIGV_DEBUG(debugging = FALSE)
}
\description{
toogles the debuging mode for AutoIGV on and off
}
