% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoIGV.R
\name{SetHeigthAndClose}
\alias{SetHeigthAndClose}
\title{SetHeigthAndClose
Sets up equal heiths to the IGV tracks}
\usage{
SetHeigthAndClose(bams.paths, file.path)
}
\arguments{
\item{bams.paths}{paths to the bams files}

\item{file.path}{path to the session file to print in}
}
\description{
SetHeigthAndClose
Sets up equal heiths to the IGV tracks
}
\author{
Maria Litovchenko
}
