% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{shift}
\alias{shift}
\title{Shift regions upstream or downstream}
\usage{
shift(gr, direction, window)
}
\arguments{
\item{gr}{regions as GenomicRanges}

\item{direction}{upstream or downstream}

\item{window}{size of shift}
}
\description{
Shift regions upstream or downstream
}
