% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulatedOutliers}
\alias{simulatedOutliers}
\title{Simulated Outliers}
\usage{
simulatedOutliers(ods, indices = F)
}
\arguments{
\item{ods}{OutriderDataSet after injection i.e. containing the metadata entry "injections"}

\item{indices}{if TRUE return the indices, else translate to feature and sample IDs}
}
\value{
return the data.table of injections
}
\description{
Conveniently retrieve the feature ID and sample IDs of the injected outliers
}
