% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{subsetExons}
\alias{subsetExons}
\title{Extract first/last exons depending on strand
takes the first hit after sorting}
\usage{
subsetExons(exon_dt, by, which = "first", strand = "both")
}
\arguments{
\item{by}{sorting criterium (transcript or gene)}

\item{dt}{containing exon info}
}
\description{
Extract first/last exons depending on strand
takes the first hit after sorting
}
