% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarise_counts}
\alias{summarise_counts}
\title{convenience function that adds summary statistics to given counts as data.table}
\usage{
summarise_counts(se)
}
\arguments{
\item{se}{summarized experiment}
}
\value{
a data.table containing the summarised values and annotations
}
\description{
convenience function that adds summary statistics to given counts as data.table
}
