% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{tssRegions}
\alias{tssRegions}
\title{Transcription start site regions}
\usage{
tssRegions(exon_dt, window, tx_id = "tx_id")
}
\arguments{
\item{window}{number of bp for regions}

\item{exons_dt}{data.table of GRanges object with columns "tx_id" for each
exon for merging by transcripts. Ranges must be strand-specific.}
}
\value{
GRanges object with upstream and downstream TSS regions
}
\description{
Get regions of specified length around the TSS
}
