% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{unnest_dt}
\alias{unnest_dt}
\title{Flatten data.tables with list entries}
\usage{
unnest_dt(dt, id_columns, list_columns, debug = FALSE)
}
\arguments{
\item{dt}{any object that can be coerced to data.table}

\item{id_columns}{all columns that do not contain any lists as entries}

\item{list_columns}{all columns that have list entries}

\item{debug}{print unlisting statement to be evaluated for each list_column}
}
\description{
Flatten data.tables with list entries
}
