% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{windowRegions}
\alias{windowRegions}
\title{Extract regions around TSS and PAS and returns a GRanges object}
\usage{
windowRegions(exons, window = 500)
}
\arguments{
\item{exons}{data.table containing at least the exon coordinates (for GRanges), "transcript_id" and "exon_id" for each exon. Must be strand-specific.}

\item{window}{number of bp that the regions should have}
}
\description{
Convenience function for TSS and PAS regions of same length
}
