% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{writeResults}
\alias{writeResults}
\title{Select columns and write to tsv file}
\usage{
writeResults(results, filename, type = "gene_wise")
}
\arguments{
\item{results}{results table from pipeline output}

\item{filename}{specify location}

\item{type}{output type, either feature_wise or gene_wise}
}
\description{
Select columns and write to tsv file
}
