% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{pasRegions}
\alias{pasRegions}
\title{Polyadenylation site regions}
\usage{
pasRegions(exons, window)
}
\arguments{
\item{exons}{data.table containing at least the exon coordinates (for GRanges), "transcript_id" and "exon_id" for each exon. Must be strand-specific.}

\item{window}{number of bp for regions}
}
\value{
data.table with coordinates and some metadata on the new regions
exon_id and transcript_id are necessary for joining the features with the rest of the exon annotation.
}
\description{
Get regions of specified length around the PAS
}
